+++ src/compas_notebook/geometry/__init__.py
@@ -1,0 +1,3 @@
+from .dot import Dot
+
+__all__ = ["Dot",] # under protest