+++ src/compas_notebook/scene/curveobject.py
@@ -1,16 +1,16 @@
 import pythreejs as three
 from compas.scene import GeometryObject
 
-from compas_notebook.conversions import line_to_threejs
+from compas_notebook.conversions.geometry import curve_to_threejs
 
 from .sceneobject import ThreeSceneObject
 
 
-class ThreeLineObject(ThreeSceneObject, GeometryObject):
-    """Scene object for drawing line."""
+class ThreeCurveObject(ThreeSceneObject, GeometryObject):
+    """Scene object for drawing curves."""
 
-    def draw(self)-> list[three.Line]:
-        """Draw the frame associated with the scene object.
+    def draw(self) -> list[three.Line]:
+        """Draw the curve as discretized polyline.
 
         Returns
         -------
@@ -18,7 +18,7 @@ class ThreeLineObject(ThreeSceneObject, GeometryObject):
             List of pythreejs objects created.
 
         """
-        geometry = line_to_threejs(self.geometry)
+        geometry = curve_to_threejs(self.geometry)
         line = three.Line(geometry, three.LineBasicMaterial(color=self.contrastcolor.hex))
 
         self._guids = [line]